backend/requirements.txt
# Core framework
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# ML and data processing
interpret>=0.6.0          # For EBM models
scikit-learn>=1.5.0
pandas>=2.2.3
numpy>=2.1.0

# OCR and document processing
pytesseract>=0.3.10       # Tesseract OCR wrapper
Pillow>=10.4.0            # Image processing
opencv-python>=4.10.0.84  # Computer vision for preprocessing
pdf2image>=1.17.0         # PDF to image conversion
PyPDF2>=3.0.1             # PDF metadata extraction for fraud detection
google-cloud-vision>=3.7.0  # Optional: Cloud Vision API

# RAG for compliance rules
faiss-cpu>=1.8.0          # Vector store
pypdf>=3.17.0             # PDF loading

# API and utilities (for future integration)
pydantic>=2.7.1
python-dotenv>=1.0.0
tenacity>=8.2.3           # Retry logic for API calls
Faker>=24.0.0             # Generate realistic mock data

# Flask API
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
Werkzeug>=3.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
